20-02-19

+++ ACPC10A NA/main.cpp
@@ -1,26 +1,25 @@
 #include <iostream>
 #include<cmath>
 using namespace std;
-// Not Accepted : Wrong Answer error , Problem code 7974
 int main()
 {
-    float a,b,c,d;
+    int a,b,c,d;
     while(true){
         cin>>a>>b>>c;
         if(a==0 && b==0 && c==0)
             break;
         else {
-                int w = b/a;
-                int z = c/b;
-            if((w)==(z)){
-                d=a*pow(z,3);
+                double w = static_cast<double>(b)/static_cast<double>(a);
+                double z = static_cast<double>(c)/static_cast<double>(b);
+                if((w)==(z)){
+                d = a*pow(z,3);
                 cout<<"GP "<<d<<endl;
             }
             else if((c-b)==(b-a)){
                 d=a+3*(c-b);
                 cout<<"AP "<<d<<endl;
             }
-        }
+          }
     }
     return 0;
 }
